requirements.txt
# python3 -m pip install -r requirements.txt

google-cloud-language
argparse
pandas
python-arango
python-dotenv
networkx  # for graph manipulation if needed